[inotify] fix remove media action

+++ deejayd/mediadb/library.py
@@ -442,9 +442,11 @@ class _Library(SignalingComponent, JSONRpcComponent):
     def remove_media(self, path, name):
         log.debug('library: removing file %s from db'\
                 % os.path.join(path, name))
-        dir_id, file_id, lm = self.lib_obj.is_file_exist(path, name)
-        self.lib_obj.remove_file(file_id)
-        self.dispatch_mupdate(file_id, 'remove')
+        rs = self.lib_obj.is_file_exist(path, name)
+        if rs is not None:
+            dir_id, file_id, lm = rs
+            self.lib_obj.remove_file(file_id)
+            self.dispatch_mupdate(file_id, 'remove')
 
     def crawl_directory(self, path, name):
         dir_path = os.path.join(path, name)
